+++ install.js
@@ -41,7 +41,7 @@ async function download() {
     process.env.PUPPETEER_DOWNLOAD_HOST ||
     process.env.npm_config_puppeteer_download_host ||
     process.env.npm_package_config_puppeteer_download_host;
-  const puppeteer = require('./index');
+  const puppeteer = require('.');
   const product =
     process.env.PUPPETEER_PRODUCT ||
     process.env.npm_config_puppeteer_product ||
