+++ package.json
@@ -2,7 +2,7 @@
   "name": "puppeteer",
   "version": "4.0.0-post",
   "description": "A high-level API to control headless Chrome over the DevTools Protocol",
-  "main": "index.js",
+  "main": "lib/index.js",
   "repository": "github:puppeteer/puppeteer",
   "engines": {
     "node": ">=10.18.1"
